Refactor move and rename
