+++ ironflow/ironflow/boxes/__init__.py
@@ -1,3 +1,4 @@
 from ironflow.ironflow.boxes.node_interface import NodeController, NodePresenter
-from ironflow.ironflow.boxes.text_output import TextOut, TextIn
+from ironflow.ironflow.boxes.text_output import TextOut
+from ironflow.ironflow.boxes.user_input import UserInput
 from ironflow.ironflow.boxes.toolbar import Toolbar
